requirement.txt
Flask-RESTful==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.8
python-dotenv==0.20.0
flask-jwt-extended
